aws_lambda_script/requirements.txt
Flask==3.0.3
boto3==1.35.33
botocore==1.35.38
flask-restx==1.3.0
aws-wsgi==0.2.7
pyjwt>=2.7.0
requests==2.32.3
cryptography==43.0.1